515-Find-Largest-Value-in-Each-Tree-Row/515-Find-Largest-Value-in-Each-Tree-Row_test.go
package leetcode

import (
	"reflect"
	"testing"
)

func Test_largestValues(t *testing.T) {
	type args struct {
		root *TreeNode
	}
	tests := []struct {
		name string
		args args
		want []int
	}{
		{
			name: "example1",
			args: args{root: &TreeNode{
				Val: 1,
				Left: &TreeNode{
					Val: 3,
					Left: &TreeNode{
						Val: 5,
					},
					Right: &TreeNode{
						Val: 3,
					},
				},
				Right: &TreeNode{
					Val: 2,
					Right: &TreeNode{
						Val: 9,
					},
				},
			}},
			want: []int{1, 3, 9},
		},
		{
			name: "example2",
			args: args{root: &TreeNode{
				Val: 1,
				Left: &TreeNode{
					Val: 2,
				},
				Right: &TreeNode{
					Val: 3,
				},
			}},
			want: []int{1, 3},
		},
		{
			name: "single node",
			args: args{root: &TreeNode{
				Val: 1,
			}},
			want: []int{1},
		},
		{
			name: "empty tree",
			args: args{root: nil},
			want: []int{},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := largestValues(tt.args.root); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("largestValues() = %v, want %v", got, tt.want)
			}
		})
	}
}
